fastapi_auth/requirements.txt
aiohttp==3.9.1
alembic==1.7.7
alembic-autogenerate-enums==0.1.2
asyncpg==0.28.0
cachetools==5.3.2
elasticsearch[async]==8.17.0
fastapi[all]===0.111.0
fastapi-cache2==0.2.1
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
redis==5.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
black==22.3.0